Move geocode and forecast data fetching into service objects
